Change type of .spec.waitReadySeconds to uint16

+++ pkg/apis/dynatrace/v1alpha1/defaults.go
@@ -2,7 +2,7 @@ package v1alpha1
 
 func SetDefaults_OneAgentSpec(obj *OneAgentSpec) {
 	if obj.WaitReadySeconds == nil {
-		obj.WaitReadySeconds = new(int32)
+		obj.WaitReadySeconds = new(uint16)
 		*obj.WaitReadySeconds = 300
 	}
 }
