+++ pkg/stub/handler.go
@@ -24,7 +24,7 @@ import (
 )
 
 // time between consecutive queries for a new pod to get ready
-const splayTimeSeconds = int32(10)
+const splayTimeSeconds = uint16(10)
 
 func NewHandler() handler.Handler {
 	return &Handler{}
@@ -135,7 +135,7 @@ func deletePods(cr *v1alpha1.OneAgent, pods []corev1.Pod) error {
 		labelSelector := labels.SelectorFromSet(getLabels(cr))
 		logrus.WithFields(logrus.Fields{"field-selector": fieldSelector, "label-selector": labelSelector}).Debug("query pod")
 		listOps := &metav1.ListOptions{FieldSelector: fieldSelector.String(), LabelSelector: labelSelector.String()}
-		for splay := int32(0); splay < *cr.Spec.WaitReadySeconds; splay += splayTimeSeconds {
+		for splay := uint16(0); splay < *cr.Spec.WaitReadySeconds; splay += splayTimeSeconds {
 			time.Sleep(time.Duration(splayTimeSeconds) * time.Second)
 			pList := getPodList()
 			err = query.List(cr.Namespace, pList, query.WithListOptions(listOps))
